Drop unused methods from order/utils.py
